+++ src/main/java/pt/sharespot/iot/core/sensor/data/IlluminanceDataDTO.java
@@ -2,9 +2,9 @@ package pt.sharespot.iot.core.sensor.data;
 
 public class IlluminanceDataDTO implements DataTypeDTO {
 
-    public Double lux;
+    public Float lux;
 
-    public static IlluminanceDataDTO of(Double lux) {
+    public static IlluminanceDataDTO of(Float lux) {
         var dataDTO = new IlluminanceDataDTO();
         dataDTO.lux = lux;
         return dataDTO;
