+++ src/main/java/pt/sharespot/iot/core/sensor/data/TemperatureDataDTO.java
@@ -2,9 +2,9 @@ package pt.sharespot.iot.core.sensor.data;
 
 public class TemperatureDataDTO implements DataTypeDTO {
 
-    public Double celsius;
+    public Float celsius;
 
-    public static TemperatureDataDTO of(Double celsius) {
+    public static TemperatureDataDTO of(Float celsius) {
         var tempCDataDTO = new TemperatureDataDTO();
         tempCDataDTO.celsius = celsius;
         return tempCDataDTO;
